I´ve refactored CharactersListView
