+++ SNK/SNK/Characters/Views/CharactersScreens/CharactersListView.swift
@@ -10,13 +10,10 @@ import CoreData
 
 struct CharactersListView: View {
     @Environment(\.colorScheme) private var colorScheme
-    @Environment(\.dismiss) private var dismiss
     @EnvironmentObject private var viewModel: CharactersViewModel
     @State private var isMenuOpen: Bool = false
     @State private var isPresented: Bool = false
     @State private var isFiltered: Bool = false
-    @State private var name: String = ""
-    @State private var status: Status = .none
     @State private var characters: [Characters]
     private var episodes: [Episodes]
     @State private var originalCharacters: [Characters]
@@ -43,54 +40,12 @@ struct CharactersListView: View {
                                     CharacterDetailView(character: item, characters: self.originalCharacters, episodes: self.episodes)
                                 } label: {
                                     VStack {
-                                        HStack {
-                                            ImageView(imageData: item.img,
-                                                      imageWidth: 125,
-                                                      imageHeight: 125)
-                                            VStack {
-                                                if let name = item.name, name != "" {
-                                                    Text(self.localization.name_charlist_view(name: name))
-                                                        .frame(maxWidth: .infinity, alignment: .leading)
-                                                        .font(.headline)
-                                                        .multilineTextAlignment(.leading)
-                                                        .foregroundStyle(self.colorByColorScheme)
-                                                        .padding(.trailing, 15)
-                                                }
-                                                if item.status == Status.alive.rawValue && item.age > 0 {
-                                                    Text(self.localization.age_charlist_view(age: String(item.age)))
-                                                        .frame(maxWidth: .infinity, alignment: .leading)
-                                                        .font(.subheadline)
-                                                        .foregroundStyle(self.colorByColorScheme)
-                                                }
-                                                HStack {
-                                                    if let statusString = item.status, statusString != "" {
-                                                        Text(self.localization.status_charlist_view(status: statusString))
-                                                            .foregroundStyle(self.colorByColorScheme)
-                                                        let status = Status(rawValue: statusString) ?? .none
-                                                        self.getIconByStatus(status: status)
-                                                        Spacer()
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        .padding()
-                                        .accessibilityRemoveTraits(.isSelected)
+                                        CharacterCellView(character: item, localization: self.localization)
+                                            .padding()
+                                            .accessibilityRemoveTraits(.isSelected)
                                         
                                         if item == self.characters.last && item.id < 201 && !self.isFiltered {
-                                            Button {
-                                                Task {
-                                                    await self.viewModel.fetchCharacters()
-                                                }
-                                                self.characters = self.viewModel.characters
-                                                self.originalCharacters = self.viewModel.characters
-                                            } label: {
-                                                HStack {
-                                                    Image(systemName: "arrow.2.circlepath")
-                                                    Text(self.localization.load_more_data)
-                                                }
-                                            }
-                                            .font(.title3)
-                                            .foregroundStyle(self.colorByColorScheme)
+                                            self.loadMoreButton
                                         }
                                     }
                                 }
@@ -102,10 +57,8 @@ struct CharactersListView: View {
                 .blur(radius: self.isMenuOpen ? 3 : 0)
                 .sheet(isPresented: $isPresented, content: {
                     FilterView(isCharacter: true) { name, status, _ in
-                        self.name = name
-                        self.status = status
                         self.isFiltered = name.isEmpty && status == .none ? false : true
-                        self.characters = self.viewModel.filterCharacters(filterName: self.name, filterStatus: self.status)
+                        self.characters = self.viewModel.filterCharacters(filterName: name, filterStatus: status)
                     }
                 })
                 .navigationBarBackButtonHidden()
@@ -164,10 +117,6 @@ struct CharactersListView: View {
     }
 }
 
-#Preview {
-    CharactersListView(characters: [], episodes: [])
-}
-
 extension CharactersListView {
     private var colorByColorScheme: Color {
         self.colorScheme == .dark ? .white : .black
@@ -177,20 +126,20 @@ extension CharactersListView {
         DefaultLocalization()
     }
     
-    func getIconByStatus(status: Status) -> some View {
-        switch status {
-        case .alive:
-            return Image(systemName: "checkmark.circle.fill")
-                .foregroundStyle(.green)
-        case .deceased:
-            return Image(systemName: "xmark.circle.fill")
-                .foregroundStyle(.red)
-        case .unknown:
-            return Image(systemName: "exclamationmark.circle.fill")
-                .foregroundStyle(.yellow)
-        case .none:
-            return Image(systemName: "questionmark.circle.dashed")
-                .foregroundStyle(.gray)
+    private var loadMoreButton: some View {
+        Button {
+            Task {
+                await self.viewModel.fetchCharacters()
+            }
+            self.characters = self.viewModel.characters
+            self.originalCharacters = self.viewModel.characters
+        } label: {
+            HStack {
+                Image(systemName: "arrow.2.circlepath")
+                Text(self.localization.load_more_data)
+            }
         }
+        .font(.title3)
+        .foregroundStyle(self.colorByColorScheme)
     }
 }
